Use SciMLTesting v1.2 folder-based run_tests
Replace the hand-written GROUP dispatcher with SciMLTesting's run_tests.
Core (basic_functionality / algorithm_traits / bigfloat tests) runs in the main
test env for All/Core; JET (static analysis) runs in its own test/JET sub-env
only for GROUP=JET. test_groups.toml is unchanged; the curated all=["Core"] keeps
JET out of "All", preserving the original per-GROUP test set exactly.
