+++ test/JET/Project.toml
@@ -3,6 +3,8 @@ GlobalDiffEq = "1d72d19b-84cc-4cb7-a099-7cbdb9ccc67c"
 JET = "c3a54625-cd67-489e-a8e7-0a5a0ff4e31b"
 OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
 OrdinaryDiffEqSSPRK = "669c94d9-1f4b-4b64-b377-1aa079aa2388"
+SafeTestsets = "1e83bf80-4336-4d27-bf5d-d5a4f845583c"
+SciMLTesting = "09d9d899-5365-40a9-917a-5f67fddea283"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
@@ -10,3 +12,5 @@ GlobalDiffEq = "1"
 JET = "0.9, 0.10, 0.11"
 OrdinaryDiffEq = "5.42, 6.103, 7"
 OrdinaryDiffEqSSPRK = "1, 2"
+SafeTestsets = "0.0.1, 0.1"
+SciMLTesting = "1"
