+++ test/runtests.jl
@@ -1,77 +1,33 @@
-using Pkg
-using GlobalDiffEq, OrdinaryDiffEq, LinearAlgebra
-using OrdinaryDiffEqSSPRK
+using SafeTestsets
 using Test
-import DiffEqBase: SciMLBase
+using SciMLTesting
 
-const GROUP = get(ENV, "GROUP", "All")
-
-function activate_jet_env()
-    Pkg.activate("jet")
-    Pkg.develop(PackageSpec(path = dirname(@__DIR__)))
-    return Pkg.instantiate()
-end
-
-@testset "GlobalDiffEq.jl" begin
-    if GROUP == "All" || GROUP == "Core"
-        @testset "Basic functionality" begin
-            l = 1.0                             # length [m]
-            m = 1.0                             # mass[m]
-            g = 9.81                            # gravitational acceleration [m/s²]
-
-            function pendulum!(du, u, p, t)
-                du[1] = u[2]                    # θ'(t) = ω(t)
-                return du[2] = -3g / (2l) * sin(u[1]) + 3 / (m * l^2) * p(t) # ω'(t) = -3g/(2l) sin θ(t) + 3/(ml^2)M(t)
-            end
-
-            θ₀ = 0.01                           # initial angular deflection [rad]
-            ω₀ = 0.0                            # initial angular velocity [rad/s]
-            u₀ = [θ₀, ω₀]                       # initial state vector
-            tspan = (0.0, 10.0)                  # time interval
-
-            M = t -> 0.1sin(t)                    # external torque [Nm]
-
-            prob = ODEProblem(pendulum!, u₀, tspan, M)
-
-            v0 = solve(prob, Tsit5(), dt = 0.1, reltol = 1.0e-12, abstol = 0).(1:10)
-            ve = solve(
-                prob, GlobalRichardson(SSPRK33()), dt = 0.2, reltol = 1.0e-12, abstol = 0
-            ).(1:10)
-            @test norm(ve - v0) / norm(v0) < 1.0e-10
+run_tests(;
+    core = function ()
+        @safetestset "Basic functionality" begin
+            include("basic_functionality_tests.jl")
         end
-
-        @testset "Algorithm traits forwarding" begin
-            alg_inner = SSPRK33()
-            alg = GlobalRichardson(alg_inner)
-
-            @test SciMLBase.allows_arbitrary_number_types(alg) ==
-                SciMLBase.allows_arbitrary_number_types(alg_inner)
-            @test SciMLBase.allowscomplex(alg) == SciMLBase.allowscomplex(alg_inner)
-            @test SciMLBase.isautodifferentiable(alg) == SciMLBase.isautodifferentiable(alg_inner)
+        @safetestset "Algorithm traits forwarding" begin
+            include("algorithm_traits_tests.jl")
         end
-
-        @testset "BigFloat support" begin
-            function f_bf!(du, u, p, t)
-                du[1] = -u[1]
-            end
-
-            u0_bf = BigFloat[1.0]
-            tspan_bf = (BigFloat(0.0), BigFloat(1.0))
-            prob_bf = ODEProblem(f_bf!, u0_bf, tspan_bf)
-
-            sol_bf = solve(
-                prob_bf, GlobalRichardson(SSPRK33()),
-                dt = BigFloat(0.1), reltol = BigFloat(1.0e-3), abstol = BigFloat(1.0e-6)
-            )
-
-            @test eltype(sol_bf.u[end]) == BigFloat
-            # Check solution is reasonable (e^-1 ≈ 0.368)
-            @test isapprox(sol_bf.u[end][1], exp(BigFloat(-1)), rtol = 1.0e-4)
+        return @safetestset "BigFloat support" begin
+            include("bigfloat_tests.jl")
         end
-    end
-
-    if GROUP == "JET"
-        activate_jet_env()
-        @testset "JET static analysis" include("jet/jet_tests.jl")
-    end
-end
+    end,
+    groups = Dict(
+        # JET runs the static-analysis tests in its own environment. The original
+        # dispatcher ran these only for GROUP=="JET" (never as part of "All"), so JET
+        # is an env-bearing group kept out of the curated `all`.
+        "JET" => (;
+            env = joinpath(@__DIR__, "JET"),
+            body = function ()
+                return @safetestset "JET static analysis" begin
+                    include(joinpath(@__DIR__, "JET", "jet_tests.jl"))
+                end
+            end,
+        ),
+    ),
+    # The original runtests.jl ran the Core body for GROUP=All and GROUP=Core, and
+    # ran JET only for GROUP=JET (never under "All"). Curate "All" to Core only.
+    all = ["Core"],
+)
